+++ packages/device_info/android/src/main/java/io/flutter/plugins/deviceinfo/DeviceInfoPlugin.java
@@ -23,8 +23,7 @@ public class DeviceInfoPlugin implements FlutterPlugin {
 
   @Override
   public void onAttachedToEngine(FlutterPlugin.FlutterPluginBinding binding) {
-    setupMethodChannel(
-        binding.getFlutterEngine().getDartExecutor(), binding.getApplicationContext());
+    setupMethodChannel(binding.getBinaryMessenger(), binding.getApplicationContext());
   }
 
   @Override